+++ src/emulators/index.ts
@@ -2,3 +2,4 @@ export * from './firebase-storage.js';
 export * from './firestore.js';
 export * from './identity-platform.js';
 export * from './pubsub.js';
+export * from './spanner.js';