+++ src/functions/index.ts
@@ -3,10 +3,12 @@ import { EmulatorStartForFirebaseStorage } from './emulator-start-firebase-stora
 import { EmulatorStartForFirestore } from './emulator-start-firestore.js';
 import { EmulatorStartForIdentityPlatform } from './emulator-start-identity-platform.js';
 import { EmulatorStartForPubSub } from './emulator-start-pubsub.js';
+import { EmulatorStartForSpanner } from './emulator-start-spanner.js';
 import { EmulatorStopForFirebaseStorage } from './emulator-stop-firebase-storage.js';
 import { EmulatorStopForFirestore } from './emulator-stop-firestore.js';
 import { EmulatorStopForIdentityPlatform } from './emulator-stop-identity-platform.js';
 import { EmulatorStopForPubSub } from './emulator-stop-pubsub.js';
+import { EmulatorStopForSpanner } from './emulator-stop-spanner.js';
 import { GoogleFirebaseStorageMergeRules } from './google-firebase-storage-merge-rules.js';
 import { GoogleFirestoreMergeRules } from './google-firestore-merge-rules.js';
 import { GoogleServicesEnable } from './google-services-enable.js';
@@ -20,10 +22,12 @@ export function registerFunctions(context: ModuleRegistrationContext) {
     EmulatorStartForFirestore,
     EmulatorStartForIdentityPlatform,
     EmulatorStartForPubSub,
+    EmulatorStartForSpanner,
     EmulatorStopForFirebaseStorage,
     EmulatorStopForFirestore,
     EmulatorStopForIdentityPlatform,
     EmulatorStopForPubSub,
+    EmulatorStopForSpanner,
     GoogleFirebaseStorageMergeRules,
     GoogleFirestoreMergeRules,
     GoogleServicesEnable,